add exist check

+++ Vagrantfile
@@ -2,28 +2,66 @@
 # vi: set ft=ruby :
 
 $install_git = <<INSTALL_GIT
-  echo provisioning git...
-  apt-get install git -y
+  if ! dpkg -l git; then
+    echo provisioning git...
+    apt-get install git -y
+  else
+    echo skip provision git...
+  fi
 INSTALL_GIT
 
 $install_curl = <<INSTALL_Curl
-  echo provisioning curl...
-  apt-get install curl -y
+  if ! dpkg -l curl; then
+    echo provisioning curl...
+    apt-get install curl -y
+  else
+    echo skip provision curl...
+  fi
 INSTALL_Curl
 
 $install_vim = <<INSTALL_VIM
-  echo provisioning vim...
-  apt-get install vim -y
-  curl -o /home/vagrant/.vimrc https://raw.githubusercontent.com/TongChang/dotfiles/master/vim/_vimrc
+  if ! dpkg -l vim; then
+    echo provisioning vim...
+    apt-get install vim -y
+  else
+    echo skip provision vim...
+  fi
+
+  if [ ! -f /home/vagrant/.vimrc ]; then
+    echo downloading vimrc...
+    curl -o /home/vagrant/.vimrc https://raw.githubusercontent.com/TongChang/dotfiles/master/vim/_vimrc
+  else
+    echo skip download vimrc...
+  fi
 INSTALL_VIM
 
 $install_ansible = <<INSTAL_Ansible
-  echo provisioning ansible...
-  apt-get ansible
+  if ! dpkg -l ansible; then
+    echo provisioning ansible...
+    apt-get install ansible -y
+  else
+    echo skip provision ansible...
+  fi
 INSTAL_Ansible
+
+$create_id_rsa = <<CREATE_id_rsa
+  if [ ! -f /home/vagrant/.ssh/id_rsa ]; then
+    echo create id_rsa...
+    ssh-keygen -t rsa -N bebop -C bebop -f /home/vagrant/.ssh/id_rsa
+    scp /home/vagrant/.ssh/id_rsa.pub 192.168.33.12:/home/vagrant/.ssh/
+  else
+    echo skip create id_rsa...
+  fi
+CREATE_id_rsa
+
 Vagrant.configure(2) do |config|
 
   config.vm.box = "ubuntu/trusty32"
+
+  config.vm.define :rpi_1 do |node|
+    node.vm.network :forwarded_port, guest: 22, host: 2011, id: "ssh"
+    node.vm.network :private_network, ip: "192.168.33.12"
+  end
 
   config.vm.define :manager do |node|
     node.vm.network :forwarded_port, guest: 22, host: 2001, id: "ssh"
@@ -36,12 +74,14 @@ Vagrant.configure(2) do |config|
     node.vm.provision "manager-vim", type: "shell" do |s|
       s.inline = $install_vim
     end
-  end
 
-  config.vm.define :rpi_1 do |node|
-    node.vm.network :forwarded_port, guest: 22, host: 2011, id: "ssh"
-    node.vm.network :forwarded_port, guest: 80, host: 8000, id: "http"
-    node.vm.network :private_network, ip: "192.168.33.12"
+    node.vm.provision "manager-ansible", type: "shell" do |s|
+      s.inline = $install_ansible
+    end
+
+    node.vm.provision "manager-id_rsa", type: "shell" do |s|
+      s.inline = $create_id_rsa
+    end
   end
 
 end
